third_party/blink/web_tests/fast/xpath/detached-subtree-invalidate-iterator-expected.txt

Test that iterators are invalidated even if the original context is detached.

PASS result.invalidIteratorState is false
PASS result.iterateNext().tagName is 'span'
Modifying the document...
PASS result.invalidIteratorState is true
PASS result.iterateNext() threw exception InvalidStateError: Failed to execute 'iterateNext' on 'XPathResult': The document has mutated since the result was returned..
PASS successfullyParsed is true

TEST COMPLETE
